symlink the entrypoint script to the 7.9 location
The stable SonarQube Helm chart directly refers to the entrypoint script defined in 7.9.1's image. Since this file moved in the new approach with 8.x, the images do not work for helm users. With the added symlink, this allows the helm charts to continue to function with the new images.