Support forward-compatible env vars in 7.9 LTS
The `SONARQUBE_*` environment variables used in older images have been
deprecated in favor of sticking with the `SONAR_*` equivalents (cf.
MMF-1881). This deprecation was only applied to Dockerfiles for versions
8.x, which had the side effect of making it slightly harder for the Helm
chart to support both major versions.

This change allows the non-deprecated environment variables to be read
by the 7.9 LTS image, if present, without breaking existing behavior.