Create paths for environment variables' values
By looking at the Dockerfile it is obvious that a user can override some
default paths by providing environmental variables.

However, there is no mechanism to make sure that any value provided by the user
will work. As a result, if a user overrides the `SQ_DATA_DIR` for example with
a random path outside of `/opt`, it gets an error:

```
chown: /foo4/data: No such file or directory
```

This commit, simply creates the folders for the env vars, preparing the image
for this case.